Misc clippy stuff

+++ src/movepick.rs
@@ -523,7 +523,7 @@ impl MovePickerPC {
             list: [ExtMove {
                 m: Move::NONE,
                 value: 0,
-            }; MAX_MOVES as usize],
+            }; MAX_MOVES],
         }
     }
